Move test to test_search.py
